style: test-utils renamed
