+++ frontend/naengpa/src/containers/RecipeList/CreateRecipe/CreateRecipe.test.tsx
@@ -4,7 +4,7 @@ import { mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import configureStore from 'redux-mock-store';
-import waitForComponentToPaint from '../../../../utils/waitForComponentToPaint';
+import waitForComponentToPaint from '../../../../test-utils/waitForComponentToPaint';
 import CreateRecipe from './CreateRecipe';
 import '@testing-library/jest-dom';
 import * as recipeActionCreators from '../../../store/actions/recipe';
